Switch from Float to Double when parsing Gerber data
We are losing some precision when parsing coordinate data from Gerber
and storing it in Float so we switch to Double to avoid this. A better
approach would probably have been a fixed precision representation but
Double gets rid of most of the precision loss while not significantly
changing the rest of the code.

Here is a hand wavy rough justification. `FSLAX45Y45` in MM means
coordinates have 4 integer + 5 decimal digits = 9 significant digits
max. The largest value is 9999.99999 mm.
- Double provides ~15.9 decimal digits of precision — more than enough for 9 digits. No loss.
- Float provides ~7.2 decimal digits — not enough for 9 digits. This was the bug.
- `Y14286369` decodes to 142.86369 mm (9 significant digits):
  - As Double: 142.86369 stored exactly
  - As Float: 142.86369 rounds to 142.8637 — the last 2 digits are lost

+++ gerber/lib/Gerber/Attribute.hs
@@ -46,7 +46,7 @@ data ApertureAttribute
 
 
 data Tolerance = Tolerance
-  { plusTolerance :: !Float
-  , minusTolerance :: !Float
+  { plusTolerance :: !Double
+  , minusTolerance :: !Double
   }
   deriving (Eq, Show)
